Separate mongo migration and Storage test files
